conf.d: Replicate `pyenv root` logic
Calling `pyenv root` causes the `pyenv` function to be called and
then go through the whole stack, doing all the things `pyenv` does
to eventually return us either the value of `$PYENV_ROOT` or
if that's empty `$HOME/.pyenv`.

By simply replicating this piece of code in Fish and not calling
`pyenv root` during the initialisation we manage to speed things
up even further.